Refactor analyzers to modularize context and patterns
Moved class context, code patterns, entity match, and collection task/result classes to dedicated files under models/analyzer and analyzers directories. Removed parallel and isolate-based code entity and usage collection from analyzers, simplifying the logic to sequential processing. This improves code organization, maintainability, and separation of concerns.
